better wrapping of joins as annotation iterators in Query
When wrapping a join as an AnnoIt the old implementation tried to keep track which is the "original" side and only called "next" on this one. This was done in order to avoid calling next() for underlying joins too often for overlapping joins. This approach did not work well for complex query execution plans. Instead the JoinWrapIterator now knows it's counterpart checks if next() of the underlying join has to be called because none of the sides have been called yet.