Only delete the DB object from the corpus cache after deleting the files was successful.
The deleteCorpus() function will first get a lock on the DB instance, thus even if other threads access them from the disk they won't be able to use it until deleting the file contents and the entry from the cache are complete.

Thus queries that are started just after the deletion began will still be able to finish their query but will then give free the last instance of the DB.
If someone calls "importCorpus" for the same name just after the deleting is started the lock on the same DB instance will delay the saving of the new version of the corpus until deletion is complete. Thus the corpus object will not longer be in the cache, but the consistent new version will be stored on disk.
Similar effects will occur when someone calls applyUpdate() after the deletion was started but while the corpus files are still deleted. The update will be consistent, but it will refer to an empty corpus instead to the original one.