Node names should include the document name (and the URL specific stuff) when imported from Salt.
In graphANNIS it is perfectly legal to have edges between nodes from different documents. There are several places where it is assumed that node names are unique (e.g. the graph update). This is only true in Salt for nodes inside a document, but not for a complete corpus (the same in ANNIS). Thus to make the node name unique the document name has to be used as a prefix.