The UniqueDFS should visit all nodes, but only check that a node is outputted only once.
The search can't be aborted whenever a node has been already outputted since when using a ranged search, there can be constellations where a valid node would not be outputted if the wrong path with the wrong node is searched first. Also a add a simple test case for this.