Do not provide an inverse operator if the underlying graph storage does not have the same cost for the inverse operations.
In the future, a proper cost model for access operations on the graph storage is needed, which would allow to actually compare the normal and the inverse operator to each other. Since this comparison is not implemented yet in the planner, not using the inverse operator in this cases is the more safe way.