Add additional "annis:doc" node to legacy "meta::" search to maintain backward compability.
Since the document nodes are not the direct parents of annotation nodes anymore (the text nodes are), the @* is used instead of @1,1. This additional identical node operator is needed to limit the results to the nodes that are actual documents and not texts or toplevel corpora. The more flexible @ operator allows matching these, but is an addition to AQL and therefore the users can decide on their own if they need additional filtering.