OperatorSpec::necessary_components should have the GraphDB as argument.
Some of the operators (like Dominance) can't get the necessary components from their spec without having access to the list of all components. This is currently solved by creating the spec with the help of of the GraphDB (the new() functions) but that means the specs are not actually specs but already hold state.