Use a cache strategy instead of only allowing a maximum fixed memory size.
This allows to set a dynamic percentage of how many of the free memory can be taken and to recalculate the available free memory each time a corpus is loaded.
Before the "auto" strategy calculated this value once at the beginning and if the environment changed, the graphANNIS maximum cache size did not adapt.