Replace the normal quicksort with a variant which aborts sorting of the right half, if maximum amount of items to sort is reached.
The find query is typically limited to a certain number of matches.
Still, we have to find all instances and sort them to find the first n matches.
The old implementation sorted all found items, which was costly, this one limits itself to ensure that the first n items are sorted.