Fix non-executed test queries.
Tests are excuted in parallel in Rust and thus a thread local variable was used to store a corpus storage for each thread. Since there is now a safe-guard in place that disallows more than one corpus storage per file-folder, most of these corpus storages failed to load and the query search tests were ignored.