Remove the /local-login end-point and instead document how to create a JWT token for development.
Having our own API to exchange user-name/password combinations instead of just using very simple to install IdP solutions like Keycloak is a significant overhead. GraphANNIS is not an indenty provider but a REST service to search in corporq. The only situation where it makes sense to not have an IdP service is when developing or testing the API. We document one of the command line interfaces for this scenario.