Fail ignored search tests when test corpora are not imported.
These tests are only executed when explicitly asked for because they takes some time (especially in debug) and we can't bundle the corpora in the git repo. Thus it is OK to fail them if some external conditions are not met.
It is better the dev needs to download the test corpora than a test case is accidentally skipped on a regular basis e.g in CI (which did happen before).