Directly use SSTables in the GraphUpdate instead of the more expensive DiskMap.
GraphUpdate needs to be optimized for adding items in sorted order with minimal overhead.
Caching the C0 table in memory is not really helpful in this case, and merging C0 and C1 will get expensive.
Instead, GraphUpdate directly creates a Table with the TableBuilder, since all input is already sorted.