Do not estimate size with graphannis-malloc_size_of.
The corpus storage now just measures the total memory used for the process and removes corpora from the cache until the threshold is reached.
The DiskMap implementation only implements a fixed number of entries as eviction strategy.

Using graphannis-malloc_size_of meant we were bound to the sytems allocator and e.g. could not change it to jemalloc in case we want fix memory fragmentation issues.